fixed platform ccast issue

+++ programs/util.c
@@ -510,7 +510,8 @@ void Stats(double start, int blockSize)
     int64_t compBlocks = blocks;
     double total = CurrTime() - start, mbs;
 
-    printf("took %6.3f seconds, blocks = %ld\n", total, compBlocks);
+    printf("took %6.3f seconds, blocks = %llu\n", total,
+           (unsigned long long)compBlocks);
 
     mbs = compBlocks * blockSize / MEGABYTE / total;
     printf("MB/s = %8.1f\n", mbs);
